A couple more methods are moved from Tiles class to underlying classes.

+++ src/Tiles/Caching/PrefetchManager.cpp
@@ -20,9 +20,6 @@
 
 #include "stdafx.h"
 #include "PrefetchManager.h"
-#include "BaseProvider.h"
-#include "TileCacheManager.h"
-#include "ITileLoader.h"
 
 ::CCriticalSection PrefetchManagerFactory::_lock;
 vector<PrefetchManager*> PrefetchManagerFactory::_managers;
@@ -69,10 +66,10 @@ void PrefetchManager::BuildDownloadList(BaseProvider* provider, int zoom, CRect 
 	provider->get_Projection()->GetTileMatrixMinXY(zoom, size1);
 	provider->get_Projection()->GetTileMatrixMaxXY(zoom, size2);
 
-	double minX = (int)BaseProjection::Clip(indices.left, size1.cx, size2.cx);
-	double maxX = (int)BaseProjection::Clip(indices.right, size1.cy, size2.cy);
-	double minY = (int)BaseProjection::Clip(MIN(indices.top, indices.bottom), size1.cx, size2.cx);
-	double maxY = (int)BaseProjection::Clip(MAX(indices.top, indices.bottom), size1.cy, size2.cy);
+	int minX = (int)BaseProjection::Clip(indices.left, size1.cx, size2.cx);
+	int maxX = (int)BaseProjection::Clip(indices.right, size1.cy, size2.cy);
+	int minY = (int)BaseProjection::Clip(MIN(indices.top, indices.bottom), size1.cx, size2.cx);
+	int maxY = (int)BaseProjection::Clip(MAX(indices.top, indices.bottom), size1.cy, size2.cy);
 
 	int centX = (maxX + minX) / 2;
 	int centY = (maxY + minY) / 2;
